py-pydantic-lab/requirements.txt
pydantic-ai
logfire
pydantic
python-dotenv
httpx
azure-identity
openai
mcp
anyio
